requirements.txt
attrs==23.2.0
blinker==1.8.2
camelcase==0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.3
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
platformdirs==4.2.0
pycodestyle==2.11.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.19.0
six==1.16.0
urllib3==2.2.0
virtualenv==20.25.1
gunicorn==20.1.0
Werkzeug==3.0.3
sqlalchemy
passlib==1.7.4
mysqlclient==2.1.0
flask_cors
firebase-admin
firebase
flask_jwt_extended
flask_bcrypt
cerberus